+++ CHANGES.rst
@@ -4,8 +4,10 @@ Changelog
 0.3 (unreleased)
 ----------------
 
-- Nothing changed yet.
-
+- Make sure overrided vocabulary `plone.app.contenttypes.metadatafields` is
+  also used when adding a new DashboardCollection, so when current context is
+  not a DashboardCollection but the parent.
+  [gbastien]
 
 0.2 (2018-09-04)
 ----------------